move example to test file
