+++ pkg/codec/byte_string_modifier.go
@@ -2,7 +2,6 @@ package codec
 
 import (
 	"bytes"
-	"crypto/rand"
 	"encoding/base64"
 	"errors"
 	"fmt"
@@ -41,37 +40,6 @@ func (m *ExampleAddressModifier) DecodeAddress(str string) ([]byte, error) {
 
 func (m *ExampleAddressModifier) Length() int {
 	return 32
-}
-
-func ExampleAddressBytesToStringModifier() {
-	type onChainNested struct {
-		X []byte
-	}
-
-	type onChain struct {
-		A [32]byte
-		B onChainNested
-	}
-
-	encoder := &ExampleAddressModifier{}
-	mod := NewPathTraverseAddressBytesToStringModifier([]string{"B.X"}, encoder, true)
-
-	// call RetypeToOffChain first with empty itemType to set base types
-	offChainType, _ := mod.RetypeToOffChain(reflect.TypeOf(&onChain{}), "")
-
-	fmt.Println("offChainType:")
-	fmt.Println(offChainType)
-	// offChainType:
-	// struct { A: string; B: struct { X: string } }
-
-	// calls to transform can transform the entire struct or nested fields specified by itemType
-	onChainAddress := [32]byte{}
-	_, _ = rand.Read(onChainAddress[:])
-
-	offChainAddress, _ := mod.TransformToOffChain(onChainAddress, "A")
-
-	// the onChainAddress value is modified to the offChainType
-	fmt.Println(offChainAddress)
 }
 
 // AddressModifier defines the interface for encoding, decoding, and handling addresses.
